Move LayerTree to tqec.compile
