+++ src/tqec/compile/tree.py
